+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.5.0] unreleased
 ### Added
+- You can now use the Editor from [here](https://lbedrockl.github.io/Pixel-Quest-Level-Editor/)
 - Copy pasting now includes, volumes, objects and all layer
 - Added drown volume used for the drown animation in Lost Gifts, but will also work for custom death animation on any tileset